bin/test-snap-refresh-config-path.sh: add test for config files after refresh
An important part of checking the refreshes from previous revisions work
is that the config files that were generated once on install with values
that were specific to that revision (i.e. file paths like
/var/snap/edgexfoundry/866/...) are fixed by the post-refresh hook  to
use the "current" symlink (i.e. /var/snap/edgexfoundry/current/...) so
that one can upgrade directly from the stable or delhi channels to this
revision with working config files.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>