Fixup test-network-interfaces test
This was the only tests that didn't include a 120s
sleep to ensure all of the services had been started
before testing the necessary conditions. So basically
a race condition which meant sometimes it succeeded,
and sometimes it failed. This change also modifies
the 'lsof' command to explicitly ask for the the
status of port 48080 instead of querying all open
IPv4 network files.

Signed-off-by: Tony Espy <espy@canonical.com>